feat(tax): widen tax_rate_bp to BIGINT + align Go types (HYG-3)
Closes out the tax-rate representation item. The functional half of
HYG-3 already shipped in migration 0003, which dropped the legacy
NUMERIC(6,2) / NUMERIC(5,4) tax_rate columns so all tax arithmetic now
runs through integer basis points end-to-end — no floats anywhere in
the money path.

Remaining gap: tax_rate_bp was INT4 while every peer money column
(tax_amount_cents, subtotal_cents, total_amount_cents, …) is BIGINT.
INT4's 21-million-percent ceiling is not a real overflow risk, but the
schema inconsistency is exactly the "why is this one column different?"
trap the HYG- items exist to remove.

Migration 0014 widens tax_rate_bp on tenant_settings, invoices, and
invoice_line_items with ALTER … TYPE BIGINT (metadata-only on Postgres
12+, safe at any table size). Go-side TaxRateBP / taxRateBP fields flip
from int to int64 and TaxOverrideRateBP from *int to *int64 so the
domain types match the column width exactly.

+++ internal/billing/engine.go
@@ -147,7 +147,7 @@ func (e *Engine) SetCouponApplier(c CouponApplier) {
 // SubtotalCents - DiscountCents + TaxAmountCents = customer total.
 type TaxApplication struct {
 	TaxAmountCents int64
-	TaxRateBP      int
+	TaxRateBP      int64
 	TaxName        string
 	TaxCountry     string
 	TaxID          string
